requirements.txt
opencv-contrib-python-headless
numpy
Pillow
matplotlib
Flask
gunicorn
